backend/requirements.txt
flask
pvporcupine
pyaudio
google-generativeai
python-dotenv
flask-cors